+++ CHANGELOG.md
@@ -69,6 +69,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Core: Fixed extra allocations when processing slashings.
 - remove unneeded container in blob sidecar ssz response
 - Light client support: create finalized header based on finalizedBlock's version, not attestedBlock.
+- Light client support: fix light client attested header execution fields' wrong version bug.
 - Testing: added custom matcher for better push settings testing.
 
 ### Security